Library upgrades
FastK no longer prints histogram, only to file.
FastK accepts relative profile option, but does not implment
DATA_BLOCK: boff is an int not an int64
libfastk:
   Streams no longer cut low counts (=> Haplex, Homex,
                                                                     Vennex)
   Bug fix in ASCII sequence output
   Fetch*  index is an int64, not int
   Find_Kmer return index not count
   GoTo routines introduced (and debugged)
   Stream makes current ptr and index visible
   Logex continues to develop
   Documentation reflects all of above

+++ split.c
@@ -117,7 +117,7 @@ static void *padded_minimizer_thread(void *arg)
   int64         *count   = data->count;
 
   char          *bases   = block->bases;
-  int64         *boff    = block->boff+1;
+  int           *boff    = block->boff+1;
 
   int   P2M2 = PAD2-2;
 
@@ -993,7 +993,7 @@ static int        short_read;  //   There was at least one read < KMER (after pr
 void Distribute_Block(DATA_BLOCK *block, int tid)
 { int    nreads  = block->nreads;
   char  *bases   = block->bases;
-  int64 *boff    = block->boff+1;
+  int   *boff    = block->boff+1;
 
   Min_File *out   = ogroup[tid];
   int64     nidx  = nfirst[tid];
